+++ client/java/src/main/java/io/openlineage/client/transports/TransportConfig.java
@@ -11,7 +11,8 @@ import com.fasterxml.jackson.annotation.JsonTypeInfo;
 @JsonSubTypes({
   @JsonSubTypes.Type(value = ConsoleConfig.class, name = "console"),
   @JsonSubTypes.Type(value = HttpConfig.class, name = "http"),
-  @JsonSubTypes.Type(value = KafkaConfig.class, name = "kafka")
+  @JsonSubTypes.Type(value = KafkaConfig.class, name = "kafka"),
+  @JsonSubTypes.Type(value = KinesisConfig.class, name = "kinesis"),
 })
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, include = JsonTypeInfo.As.PROPERTY, property = "type")
 public interface TransportConfig {}