+++ integration/airflow/openlineage/airflow/extractors/extractors.py
@@ -44,6 +44,9 @@ _extractors = list(
                 'openlineage.airflow.extractors.redshift_data_extractor.RedshiftDataExtractor'
             ),
             try_import_from_string(
+                'openlineage.airflow.extractors.trino_extractor.TrinoExtractor'
+            ),
+            try_import_from_string(
                 'openlineage.airflow.extractors.athena_extractor.AthenaExtractor'
             ),
             try_import_from_string(
@@ -60,7 +63,10 @@ _extractors = list(
             ),
             try_import_from_string(
                 'openlineage.airflow.extractors.s3_extractor.S3CopyObjectExtractor'
-            )
+            ),
+            try_import_from_string(
+                'openlineage.airflow.extractors.s3_extractor.S3FileTransformExtractor'
+            ),
         ],
     )
 )
@@ -76,6 +82,7 @@ _check_providers = {
     "MySqlExtractor": "mysql",
     "BigQueryExtractor": ["gcpbigquery", "google_cloud_platform"],
     "SnowflakeExtractor": "snowflake",
+    "TrinoExtractor": "trino",
     "AthenaExtractor": "aws",
     "SFTPExtractor": ["sftp", "ssh"],
 }