+++ integration/spark/app/src/main/java/io/openlineage/spark/agent/OpenLineageSparkListener.java
@@ -6,15 +6,9 @@
 package io.openlineage.spark.agent;
 
 import static io.openlineage.spark.agent.util.ScalaConversionUtils.asJavaOptional;
-import static io.openlineage.spark.agent.util.SparkConfUtils.findSparkConfigKey;
-import static io.openlineage.spark.agent.util.SparkConfUtils.findSparkConfigKeyDouble;
-import static io.openlineage.spark.agent.util.SparkConfUtils.findSparkConfigKeysStartsWith;
-import static io.openlineage.spark.agent.util.SparkConfUtils.findSparkUrlParams;
 
 import io.openlineage.client.Environment;
 import io.openlineage.client.OpenLineage;
-import io.openlineage.client.transports.KinesisConfig;
-import io.openlineage.client.transports.TransportConfig;
 import io.openlineage.spark.agent.lifecycle.ContextFactory;
 import io.openlineage.spark.agent.lifecycle.ExecutionContext;
 import io.openlineage.spark.agent.util.ScalaConversionUtils;
@@ -60,20 +54,6 @@ public class OpenLineageSparkListener extends org.apache.spark.scheduler.SparkLi
       Collections.synchronizedMap(new HashMap<>());
   private static final Map<Integer, ExecutionContext> rddExecutionRegistry =
       Collections.synchronizedMap(new HashMap<>());
-
-  public static final String SPARK_CONF_TRANSPORT_TYPE = "openlineage.transport.type";
-  public static final String SPARK_CONF_CONSOLE_TRANSPORT = "openlineage.consoleTransport";
-  public static final String SPARK_CONF_URL_KEY = "openlineage.url";
-  public static final String SPARK_CONF_HOST_KEY = "openlineage.host";
-  public static final String SPARK_CONF_API_VERSION_KEY = "openlineage.version";
-  public static final String SPARK_CONF_NAMESPACE_KEY = "openlineage.namespace";
-  public static final String SPARK_CONF_JOB_NAME_KEY = "openlineage.parentJobName";
-  public static final String SPARK_CONF_PARENT_RUN_ID_KEY = "openlineage.parentRunId";
-  public static final String SPARK_CONF_TIMEOUT = "openlineage.timeout";
-  public static final String SPARK_CONF_API_KEY = "openlineage.apiKey";
-  public static final String SPARK_CONF_URL_PARAM_PREFIX = "openlineage.url.param";
-  private static final String SPARK_CONF_APP_NAME = "openlineage.appName";
-  private static final String SPARK_CONF_FACETS_DISABLED = "openlineage.facets.disabled";
   private static WeakHashMap<RDD<?>, Configuration> outputs = new WeakHashMap<>();
   private static ContextFactory contextFactory;
   private static JobMetricsHolder jobMetrics = JobMetricsHolder.getInstance();
@@ -294,7 +274,7 @@ public class OpenLineageSparkListener extends org.apache.spark.scheduler.SparkLi
     SparkEnv sparkEnv = SparkEnv$.MODULE$.get();
     if (sparkEnv != null) {
       try {
-        ArgumentParser args = parseConf(sparkEnv.conf());
+        ArgumentParser args = ArgumentParser.parse(sparkEnv.conf());
         contextFactory = new ContextFactory(new EventEmitter(args));
       } catch (URISyntaxException e) {
         log.error("Unable to parse open lineage endpoint. Lineage events will not be collected", e);
@@ -309,60 +289,5 @@ public class OpenLineageSparkListener extends org.apache.spark.scheduler.SparkLi
   private static boolean checkIfDisabled() {
     String isDisabled = Environment.getEnvironmentVariable("OPENLINEAGE_DISABLED");
     return Boolean.parseBoolean(isDisabled);
-  }
-
-  private void commonConfigParse(ArgumentParser.ArgumentParserBuilder builder, SparkConf conf) {
-    builder
-        .timeout(findSparkConfigKeyDouble(conf, SPARK_CONF_TIMEOUT))
-        .apiKey(findSparkConfigKey(conf, SPARK_CONF_API_KEY).filter(str -> !str.isEmpty()))
-        .appName(findSparkConfigKey(conf, SPARK_CONF_APP_NAME).filter(str -> !str.isEmpty()))
-        .urlParams(findSparkUrlParams(conf, SPARK_CONF_URL_PARAM_PREFIX));
-
-    findSparkConfigKey(conf, SPARK_CONF_FACETS_DISABLED).ifPresent(builder::disabledFacets);
-    findSparkConfigKey(conf, SPARK_CONF_HOST_KEY).ifPresent(builder::host);
-    findSparkConfigKey(conf, SPARK_CONF_API_VERSION_KEY).ifPresent(builder::version);
-    findSparkConfigKey(conf, SPARK_CONF_NAMESPACE_KEY).ifPresent(builder::namespace);
-    findSparkConfigKey(conf, SPARK_CONF_JOB_NAME_KEY).ifPresent(builder::jobName);
-    findSparkConfigKey(conf, SPARK_CONF_PARENT_RUN_ID_KEY).ifPresent(builder::parentRunId);
-  }
-
-  private ArgumentParser parseConf(SparkConf conf) {
-    Optional<String> url = findSparkConfigKey(conf, SPARK_CONF_URL_KEY);
-    Optional<String> transportType = findSparkConfigKey(conf, SPARK_CONF_TRANSPORT_TYPE);
-    ArgumentParser.ArgumentParserBuilder builder = ArgumentParser.builder();
-    commonConfigParse(builder, conf);
-
-    if (url.isPresent()) {
-      ArgumentParser.parse(builder, url.get());
-    } else if (transportType.isPresent()) {
-      // go and check the specific transport type setting
-      String mode = transportType.get().toLowerCase();
-      Optional<TransportConfig> transportConfig = Optional.empty();
-      switch (mode) {
-        case "kinesis":
-          Map<String, String> config =
-              findSparkConfigKeysStartsWith(conf, "spark.openlineage.transport.kinesis.");
-          KinesisConfig kinesisConfig = new KinesisConfig();
-          kinesisConfig.setStreamName(config.get("streamName"));
-          kinesisConfig.setRegion(config.get("region"));
-          kinesisConfig.setRoleArn(Optional.ofNullable(config.get("roleArn")));
-          Properties properties = new Properties();
-          properties.putAll(config);
-          kinesisConfig.setProperties(properties);
-          transportConfig = Optional.of(kinesisConfig);
-        default:
-          // todo, we may support other java client transport mode in this way in the future
-      }
-      builder.transportMode(Optional.of(mode)).transportConfig(transportConfig);
-    } else {
-      boolean consoleMode =
-          findSparkConfigKey(conf, SPARK_CONF_CONSOLE_TRANSPORT)
-              .map(Boolean::valueOf)
-              .filter(v -> v)
-              .orElse(false);
-      builder.consoleMode(consoleMode);
-    }
-
-    return builder.build();
   }
 }