+++ proxy/client/transports/http_transport.go
@@ -1,0 +1,10 @@
+package transports
+
+type HttpTransport struct {
+	OpenLineageURL    string
+	OpenLineageAPIKey string
+}
+
+func (transport *HttpTransport) Emit(lineageEvent string) error {
+	return nil
+}